nlp-service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic-settings>=2.2.0
transformers>=4.40.0
torch>=2.2.0
langdetect>=1.0.9
scikit-learn>=1.4.0
pandas>=2.2.0
joblib>=1.3.0
sentencepiece>=0.2.0
sacremoses>=0.1.1
